Adjust to code style
