+++ betting_params.go
@@ -13,6 +13,21 @@ type RunnerID struct {
 	Handicap    float64
 }
 
+type EXBestOffersOverrides struct {
+	BestPricesDepth          int          `json:"bestPricesDepth,omitempty"`
+	RollupModel              ERollupModel `json:"rollupModel,omitempty"`
+	RollupLimit              int          `json:"rollupLimit,omitempty"`
+	RollupLiabilityThreshold float64      `json:"rollupLiabilityThreshold,omitempty"`
+	RollupLiabilityFactor    int          `json:"rollupLiabilityFactor,omitempty"`
+}
+
+type PriceProjection struct {
+	PriceData             []EPriceData           `json:"priceData,omitempty"`
+	EXBestOffersOverrides *EXBestOffersOverrides `json:"exBestOffersOverrides,omitempty"`
+	Virtualise            bool                   `json:"virtualise,omitempty"`
+	RolloverStakes        bool                   `json:"rolloverStakes,omitempty"`
+}
+
 type Filter struct {
 	Wallet                       EWallet            `json:"wallet,omitempty"`
 	Locale                       string             `json:"locale,omitempty"`
@@ -45,19 +60,4 @@ type Filter struct {
 	TimeGranularity              ETimeGranularity   `json:"granularity,omitempty"`
 	PlaceOrdersMarketID          string             `json:"marketId,omitempty"`
 	PlaceOrdersInstructions      []PlaceInstruction `json:"instructions,omitempty"`
-}
-
-type PriceProjection struct {
-	PriceData             []EPriceData           `json:"priceData,omitempty"`
-	EXBestOffersOverrides *EXBestOffersOverrides `json:"exBestOffersOverrides,omitempty"`
-	Virtualise            bool                   `json:"virtualise,omitempty"`
-	RolloverStakes        bool                   `json:"rolloverStakes,omitempty"`
-}
-
-type EXBestOffersOverrides struct {
-	BestPricesDepth          int          `json:"bestPricesDepth,omitempty"`
-	RollupModel              ERollupModel `json:"rollupModel,omitempty"`
-	RollupLimit              int          `json:"rollupLimit,omitempty"`
-	RollupLiabilityThreshold float64      `json:"rollupLiabilityThreshold,omitempty"`
-	RollupLiabilityFactor    int          `json:"rollupLiabilityFactor,omitempty"`
 }
