+++ scores_params.go
@@ -1,5 +1,10 @@
 package betting
 
+type UpdateKey struct {
+	EventID                     string `json:"eventId,omitempty"`
+	LastUpdateSequenceProcessed int64  `json:"lastUpdateSequenceProcessed,omitempty"`
+}
+
 type ScoresParams struct {
 	EventIDs     []string       `json:"eventIds,omitempty"`
 	EventTypeIDs []string       `json:"eventTypeIds,omitempty"`
